requirements.txt
Django==3.0.14
djangorestframework==3.11.2
